+++ sentry-rails/spec/spec_helper.rb
@@ -1,4 +1,5 @@
 require "bundler/setup"
+require "debug"
 require "pry"
 
 require "sentry-ruby"